+++ src/main/resources/server-scripts/LaunchServer.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 java -version
-java -Xmx2G -XX:MaxPermSize=256M %%ARGUMENTS%% -jar %%SERVERJAR%% "$@"
+java -Xmx2G %%ARGUMENTS%% -jar %%SERVERJAR%% "$@"
 read -n1 -r -p "Press any key to close..."