+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Remove remote error reporting when pack image couldn't be loaded
+- Checking for mod update on CurseForge having issue when it no longer exists
 
 ### Misc