+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Show progress dialog when using the Install Fabric API button when adding mods to an instance
 - Show progress dialog when adding a mod from modrinth to an instance
 - Add in platform message from config to packs browser
+- Add some debug logging around instance launch preperation
 
 ### Fixes
 - Remove remote error reporting when pack image couldn't be loaded