+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.10.3
+## 3.4.10.4
 
 ### New Features
 
 ### Fixes
-- Issues with Technic packs using newer Forge versions not installing
-- Make the Update Data button on the launcher force update Minecraft and Runtime info
 
 ### Misc